# internal/delivery/http/v1/books.go
package v1

import (
	"fmt"
	"github.com/gorilla/mux"
	api_errors "github.com/plutonio00/books-api/internal/error"
	"net/http"
)

func (h *Handler) initBooksRoutes(router *mux.Router) {
	booksRouter := router.PathPrefix(h.baseEndpoint + "books").Subrouter()
	booksRouter.Use(h.authWithToken)
	booksRouter.HandleFunc("/list", h.getBooksList).Methods("GET")
	booksRouter.HandleFunc("/{id:[0-9]+}", h.getBookById).Methods("GET")
	booksRouter.HandleFunc("/update", h.updateBook).Methods("POST")
	booksRouter.HandleFunc("/delete/{id:[0-9]+}", h.deleteBook).Methods("DELETE")
}

// @Summary Books list
// @Tags books
// @Description endpoint for getting all books
// @Produce json
// @Security ApiTokenAuth
// @Success 200 {object} ApiResponse{result=[]model.Book}
// @Failure 400 {object} ApiResponse{result=string}
// @Failure 500 {object} ApiResponse{result=string}
// @Failure default {object} ApiResponse{result=string}
// @Router /books/list [get]
func (h *Handler) getBooksList(w http.ResponseWriter, r *http.Request) {
	data, err := h.services.Books.GetBooksList()

	if err != nil {
		if err == api_errors.ErrBookNotFound {
			jsonResponse(w, http.StatusNotFound, err.Error())
			return
		}

		jsonResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	jsonResponse(w, http.StatusOK, data)
}

// @Summary Books by id
// @Tags books
// @Description endpoint for getting book by id
// @Produce json
// @Security ApiTokenAuth
// @Param id path integer true "Book's id"
// @Success 200 {object} ApiResponse{result=model.Book}
// @Failure 400 {object} ApiResponse{result=string}
// @Failure 404 {object} ApiResponse{result=string}
// @Failure 500 {object} ApiResponse{result=string}
// @Failure default {object} ApiResponse{result=string}
// @Router /books/{id} [get]
func (h *Handler) getBookById(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id := vars["id"]

	book, err := h.services.Books.FindById(id)

	if err != nil {
		if book == nil {
			jsonResponse(w, http.StatusNotFound, err.Error())
			return
		}

		jsonResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	jsonResponse(w, http.StatusOK, book)
	return
}

// @Summary Update book by id
// @Tags books
// @Description endpoint for update book by id
// @Produce json
// @Security ApiTokenAuth
// @Param id formData integer true "Book's id"
// @Param title formData string false "Book's title"
// @Param release_year formData integer false "Book's release year"
// @Param author_id formData integer false "Book's author"
// @Success 200 {object} ApiResponse{result=string}
// @Failure 400 {object} ApiResponse{result=string}
// @Failure 404 {object} ApiResponse{result=string}
// @Failure 500 {object} ApiResponse{result=string}
// @Failure default {object} ApiResponse{result=string}
// @Router /books/update [post]
func (h *Handler) updateBook(w http.ResponseWriter, r *http.Request) {
	r.ParseMultipartForm(0)
	params := r.Form

	id := r.Form.Get("id")

	if id == "" {
		jsonResponse(w, http.StatusBadRequest, "Missing mandatory parameter 'id'")
		return
	}

	keys, values := h.ParsePostParamsToKeyValueSlices(params)

	values = append(values, id)
	err := h.services.Books.UpdateById(keys, values)

	if err != nil {
		if err == api_errors.ErrBookNotFound {
			jsonResponse(w, http.StatusNotFound, err.Error())
			return
		}

		jsonResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	jsonResponse(w, http.StatusOK, "Books were updated successfully")
	return
}

// @Summary Delete book by id
// @Tags books
// @Description endpoint for delete book by id
// @Produce json
// @Security ApiTokenAuth
// @Param id path integer true "Book's id"
// @Success 200 {object} ApiResponse{result=string}
// @Failure 400 {object} ApiResponse{result=string}
// @Failure 404 {object} ApiResponse{result=string}
// @Failure 500 {object} ApiResponse{result=string}
// @Failure default {object} ApiResponse{result=string}
// @Router /books/delete/{id} [delete]
func (h *Handler) deleteBook(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id := vars["id"]

	err := h.services.Books.DeleteById(id)

	if err != nil {
		if err == api_errors.ErrBookNotFound {
			jsonResponse(w, http.StatusNotFound, err.Error())
			return
		}

		jsonResponse(w, http.StatusInternalServerError, err.Error())
		return
	}

	jsonResponse(w, http.StatusOK, fmt.Sprintf("Book with id %s deleted", id))
	return
}